requirements.txt
openclaw>=2026.3
chainlit>=2.11
google-genai>=1.0
pydantic>=2.0
python-dotenv>=1.0
